+++ HaselTweaks/Tweaks/GlamourDresserArmoireAlert.cs
@@ -1,0 +1,30 @@
+using HaselTweaks.Windows;
+
+namespace HaselTweaks.Tweaks;
+
+[Tweak]
+public unsafe partial class GlamourDresserArmoireAlert : Tweak
+{
+    public override void Enable()
+    {
+        if (IsAddonOpen("MiragePrismPrismBox"))
+            Service.WindowManager.OpenWindow<GlamourDresserArmoireAlertWindow>();
+    }
+
+    public override void Disable()
+    {
+        Service.WindowManager.CloseWindow<GlamourDresserArmoireAlertWindow>();
+    }
+
+    public override void OnAddonOpen(string addonName)
+    {
+        if (addonName == "MiragePrismPrismBox")
+            Service.WindowManager.OpenWindow<GlamourDresserArmoireAlertWindow>();
+    }
+
+    public override void OnAddonClose(string addonName)
+    {
+        if (addonName == "MiragePrismPrismBox")
+            Service.WindowManager.CloseWindow<GlamourDresserArmoireAlertWindow>();
+    }
+}